+++ src/index.ts
@@ -31,7 +31,6 @@ export * from './constants/headers'
 export {
   Conversation,
   ConversationState,
-  ConversationEndpointType,
   SendMessageObject,
   StateObject,
 } from './models/conversation';
@@ -39,6 +38,10 @@ export {
 export {
   AgentMode,
 } from './models/store';
+
+export {
+  EndpointType,
+} from './models/interfaces';
 
 export {
   DEC112Specifics,