+++ CHANGELOG.md
@@ -40,7 +40,7 @@ Rejecting the initial message of a conversation leads to the conversation being 
 - Support for starting a conversation with a custom message id.
 - Support for customizing domains in SIP headers via `EmergencySpecifics`.
 - Ability to remove listeners (conversation, state, message).
-- Ability to specify endpoint type for each conversation.
+- Ability to specify endpoint type for agent and each individual conversation.
 - Ability to specify custom user agent headers.
 - Ability to override agent's default display name at `Conversation` level.
 - Interfaces for easily fetching SIP information like TO, FROM and ROUTE headers.